requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asttokens==3.0.1
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
comm==0.2.3
debugpy==1.8.20
decorator==5.2.1
diskcache==5.6.3
executing==2.2.1
fastapi==0.135.3
filelock==3.25.2
fsspec==2026.3.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.9.2
idna==3.11
ipykernel==7.2.0
ipython==9.10.1
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jupyter_client==8.8.0
jupyter_core==5.9.1
llama_cpp_python==0.3.20
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.4.2
packaging==26.0
pandas==3.0.2
parso==0.8.6
platformdirs==4.9.6
prompt_toolkit==3.0.52
psutil==7.2.2
pure_eval==0.2.3
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
pyzmq==27.1.0
regex==2026.4.4
requests==2.33.1
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.3.0
shellingham==1.5.4
six==1.17.0
stack-data==0.6.3
starlette==1.0.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.11.0
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
transformers==5.5.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
uvicorn==0.44.0
wcwidth==0.6.0
